+++ src/shapes/Object/InteractiveObject.ts
@@ -132,7 +132,11 @@ export class InteractiveFabricObject<
     if (this.noScaleCache && targetCanvas && targetCanvas._currentTransform) {
       const target = targetCanvas._currentTransform.target,
         action = targetCanvas._currentTransform.action;
-      if (this === (target as InteractiveFabricObject) && action.slice && action.slice(0, 5) === 'scale') {
+      if (
+        this === (target as InteractiveFabricObject) &&
+        action.slice &&
+        action.slice(0, 5) === 'scale'
+      ) {
         return false;
       }
     }