+++ src/util/types.ts
@@ -12,11 +12,15 @@ import type { IText } from '../shapes/itext.class';
 import type { Textbox } from '../shapes/textbox.class';
 import type { Gradient } from '../gradient/gradient.class';
 
-export const isFiller = (filler: TFiller | string | null): filler is TFiller => {
+export const isFiller = (
+  filler: TFiller | string | null
+): filler is TFiller => {
   return !!filler && (filler as TFiller).toLive !== undefined;
 };
 
-export const isSerializableFiller = (filler: TFiller | string | null): filler is TFiller => {
+export const isSerializableFiller = (
+  filler: TFiller | string | null
+): filler is TFiller => {
   return !!filler && (filler as TFiller).toObject !== undefined;
 };