+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [next]
-
+- chore(TS): Add declare in front of properties type definitions [#8574](https://github.com/fabricjs/fabric.js/pull/8574)
 - BREAKING: remove Object.stateful and Object.statefulCache [#8573](https://github.com/fabricjs/fabric.js/pull/8573)
 - fix(IText): refactor clearing context top logic of itext to align with brush pattern, using the canvas rendering cycle in order to guard from edge cases #8560
 - fix(Canvas): `_initRetinaScaling` initializaing the scaling regardless of settings in Canvas. [#8565](https://github.com/fabricjs/fabric.js/pull/8565)